LinearEquations

Documentation for LinearEquations.

LinearEquations.genLinearEquation

LinearEquations.getCoefficient

LinearEquations.vec2ρ

LinearEquations.ρ2vec

LinearEquations.genLinearEquation — Method

genLinearEquation(linearEquationName, sz, targetEquation, paramDisc, args...)

get the time evolution form of an Master equation as linear equation form

Arguments:

linearEquationName: The name of linear equation you want to generate

sz: The size of the system, number of energy levels

targetEquation: The target equation you want to linearized

paramDisc: External code to convert the array p to parameters required

args: parameters required

Examples:

julia> function _getHamiltonian(B::Float64, Ω::ComplexF64, δ::Float64, rabiRatio::Float64)
	Γ::Float64 = 2π*5.746e6  # Decay Rate of Excited State
    μ::Float64 = 2π*0.7e6
    H = zeros(3, 3) * 0im
    H[1, 1] = μ * B
    H[2, 2] = -μ * B + δ
    H[1, 3] = Ω / 2 * rabiRatio
    H[2, 3] = Ω / 2
    H[3, 1] = Ω' / 2 * rabiRatio
    H[3, 2] = Ω' / 2
    H[3, 3] = -1im/2 * Γ
    return H
end
julia> function _∂tρ(ρ::Matrix{ComplexF64}, B::Float64, Ω::Float64, δ::Float64)
	Γ::Float64 = 2π*5.746e6  # Decay Rate of Excited State
	rabiRatio::Float64 = 1
	Ωc = Ω*(1+0im)
    H = _getHamiltonian(B, Ωc, δ, rabiRatio)
    ∂tρ = -1im * (H*ρ - ρ*H')
    ∂tρ[1, 1] = ∂tρ[1, 1] + ρ[3, 3] * Γ * rabiRatio^2 / (rabiRatio^2 + 1)
    ∂tρ[2, 2] = ∂tρ[2, 2] + ρ[3, 3] * Γ / (rabiRatio^2 + 1)
    ∂tρ
end
julia> paramDisc = quote
			B0 = p[1]
			ampB = p[2]
			phaseB = p[3]
			ωB = p[4]
			prepare = p[5]
			B = (t > prepare) ? (ampB*sin(ωB*t+phaseB)+B0) : B0
			ωΩ = p[6]
			τ = p[7]
			Ωamp = p[8]
			T = 2π/ωΩ # 脉冲周期
			lowPulse = T - τ
			Ω = (t < prepare) | ((t>prepare) & ((t-prepare-τ/2)%T>lowPulse)) ? Ωamp : 0
			δ = p[9]
			end
julia> expr = genLinearEquation(:master_equation!, 3, _∂tρ, paramDisc, :B, :Ω, :δ)

LinearEquations.getCoefficient — Method

getCoefficient(sz, argSize, masterFun)

get the coefficient of an Master equation

Arguments:

coe: The coefficient dictionary.

sz: The size of the system, number of energy levels

masterFun: The master equation

Examples:

julia> function _getHamiltonian(B::Float64, Ω::ComplexF64, δ::Float64, rabiRatio::Float64)
	Γ::Float64 = 2π*5.746e6  # Decay Rate of Excited State
    μ::Float64 = 2π*0.7e6
    H = zeros(3, 3) * 0im
    H[1, 1] = μ * B
    H[2, 2] = -μ * B + δ
    H[1, 3] = Ω / 2 * rabiRatio
    H[2, 3] = Ω / 2
    H[3, 1] = Ω' / 2 * rabiRatio
    H[3, 2] = Ω' / 2
    H[3, 3] = -1im/2 * Γ
    return H
end
julia> function _∂tρ(ρ::Matrix{ComplexF64}, B::Float64, Ω::Float64, δ::Float64)
	Γ::Float64 = 2π*5.746e6  # Decay Rate of Excited State
	rabiRatio::Float64 = 1
	Ωc = Ω*(1+0im)
	H = _getHamiltonian(B, Ωc, δ, rabiRatio)
	∂tρ = -1im * (H*ρ - ρ*H')
	∂tρ[1, 1] = ∂tρ[1, 1] + ρ[3, 3] * Γ * rabiRatio^2 / (rabiRatio^2 + 1)
	∂tρ[2, 2] = ∂tρ[2, 2] + ρ[3, 3] * Γ / (rabiRatio^2 + 1)
	∂tρ
end
julia> getCoefficient(3, 3, _∂tρ)

LinearEquations.vec2ρ — Method

vec2ρ(vec::Vector{Float64})::Matrix{ComplexF64}

Reconstruct density matrix from a vector, please refer ρ2vec

Arguments

vec::Vector{Float64}: Vectorized of density matrix

Examples

julia> begin
	ρ = zeros(3, 3) * 1im
	ρ[1, 1] = 1/2
	ρ[2, 2] = 1/2
	ρ[1, 2] = 1 + 2im
	ρ[2, 1] = 1 - 2im
	ρ[3, 3] = 0
end
julia> vec = ρ2vec(ρ)
julia> ρ2 = vec2ρ(vec)
julia> ρ2 - ρ

LinearEquations.ρ2vec — Method

ρ2vec(ρ::Matrix{ComplexF64})::Vector{Float64}

Vectorize the density matrix ho, please refer vec2ρ

Arguments

ρ::Matrix{ComplexF64}: Density matrix

Examples:

julia> begin
	ρ = zeros(3, 3) * 1im
	ρ[1, 1] = 1/2
	ρ[2, 2] = 1/2
	ρ[1, 2] = 1 + 2im
	ρ[2, 1] = 1 - 2im
	ρ[3, 3] = 0
end
julia> vec = ρ2vec(ρ)